+++ src/ForTests/Models/Core/CarsCore.php
@@ -5,7 +5,7 @@ namespace Przeslijmi\Shortquery\ForTests\Models\Core;
 use Przeslijmi\Shortquery\Data\Collection;
 use Przeslijmi\Shortquery\Data\Field;
 use Przeslijmi\Shortquery\ForTests\Models\Core\CarModel;
-use Przeslijmi\Shortquery\ForTests\Models\Girls as Girls1369;
+use Przeslijmi\Shortquery\ForTests\Models\Girls as Girls2696;
 
 /**
  * ShortQuery Collection Core class for Car Model Collection.
@@ -29,7 +29,6 @@ class CarsCore extends Collection
     /**
      * Call to add children/childrens (oneOwnerGirl) to every Instance in this Collection.
      *
-     * @since  v1.0
      * @return self
      */
     public function expandOneOwnerGirl() : self
@@ -45,7 +44,7 @@ class CarsCore extends Collection
         }
 
         // Get records with those pks.
-        $toAdd = new Girls1369(...func_get_args());
+        $toAdd = new Girls2696(...func_get_args());
         $toAdd->getLogics()->addRule('pk', $pks);
         $toAdd->read();
 
@@ -58,7 +57,6 @@ class CarsCore extends Collection
     /**
      * Call to add children/childrens (oneOwnerGirlWithSnapchat) to every Instance in this Collection.
      *
-     * @since  v1.0
      * @return self
      */
     public function expandOneOwnerGirlWithSnapchat() : self
@@ -74,7 +72,7 @@ class CarsCore extends Collection
         }
 
         // Get records with those pks.
-        $toAdd = new Girls1369(...func_get_args());
+        $toAdd = new Girls2696(...func_get_args());
         $toAdd->getLogics()->addRule('pk', $pks);
         $toAdd->read();