+++ src/ForTests/Models/Core/GirlCore.php
@@ -4,8 +4,8 @@ namespace Przeslijmi\Shortquery\ForTests\Models\Core;
 
 use Przeslijmi\Shortquery\Data\Instance;
 use Przeslijmi\Shortquery\Exceptions\Data\CollectionSliceNotPossibleException;
-use Przeslijmi\Shortquery\ForTests\Models\Car as Car9828;
-use Przeslijmi\Shortquery\ForTests\Models\Cars as Cars5598;
+use Przeslijmi\Shortquery\ForTests\Models\Car as Car9520;
+use Przeslijmi\Shortquery\ForTests\Models\Cars as Cars8490;
 use Przeslijmi\Shortquery\ForTests\Models\Core\GirlModel;
 use Przeslijmi\Shortquery\ForTests\Models\Girl;
 use Przeslijmi\Shortquery\ForTests\Models\Girls;
@@ -403,12 +403,12 @@ class GirlCore extends Instance
      *
      * @return Cars
      */
-    public function getCars() : Cars5598
+    public function getCars() : Cars8490
     {
 
         // Create empty collection if there isn't any added.
         if ($this->cars === null) {
-            $this->cars = new Cars5598();
+            $this->cars = new Cars8490();
             $this->cars->getLogics()->addFromRelation($this->grabModel()->getRelationByName('cars'));
         }
 
@@ -418,14 +418,13 @@ class GirlCore extends Instance
     /**
      * Call to add children (Cars) to this Instance.
      *
-     * @since  v1.0
      * @return self
      */
     public function expandCars() : self
     {
 
         // Get records with those pks.
-        $children = new Cars5598(...func_get_args());
+        $children = new Cars8490(...func_get_args());
 
         // If we know that we need this one - read this one.
         if ($this->getPk() !== null) {
@@ -446,11 +445,11 @@ class GirlCore extends Instance
     /**
      * Adds child-Collection to Relation Collection.
      *
-     * @param Cars5598 $collection One child-Instance of child for Relation.
+     * @param Cars8490 $collection One child-Instance of child for Relation.
      *
      * @return self
      */
-    public function addCars(Cars5598 $collection) : self
+    public function addCars(Cars8490 $collection) : self
     {
 
         // Put this Instance to this Collection.
@@ -464,12 +463,12 @@ class GirlCore extends Instance
      *
      * @return Cars
      */
-    public function getFastCars() : Cars5598
+    public function getFastCars() : Cars8490
     {
 
         // Create empty collection if there isn't any added.
         if ($this->fastCars === null) {
-            $this->fastCars = new Cars5598();
+            $this->fastCars = new Cars8490();
             $this->fastCars->getLogics()->addFromRelation($this->grabModel()->getRelationByName('fastCars'));
         }
 
@@ -479,14 +478,13 @@ class GirlCore extends Instance
     /**
      * Call to add children (Cars) to this Instance.
      *
-     * @since  v1.0
      * @return self
      */
     public function expandFastCars() : self
     {
 
         // Get records with those pks.
-        $children = new Cars5598(...func_get_args());
+        $children = new Cars8490(...func_get_args());
 
         // If we know that we need this one - read this one.
         if ($this->getPk() !== null) {
@@ -507,11 +505,11 @@ class GirlCore extends Instance
     /**
      * Adds child-Collection to Relation Collection.
      *
-     * @param Cars5598 $collection One child-Instance of child for Relation.
+     * @param Cars8490 $collection One child-Instance of child for Relation.
      *
      * @return self
      */
-    public function addFastCars(Cars5598 $collection) : self
+    public function addFastCars(Cars8490 $collection) : self
     {
 
         // Put this Instance to this Collection.